+++ tests/test_plan.py
@@ -1,3 +1,6 @@
+import pytest
+
+from etlrules.exceptions import InvalidPlanError
 from etlrules.plan import Plan
 from etlrules.backends.pandas import ProjectRule, RenameRule, SortRule
 
@@ -37,3 +40,12 @@ def test_plan_to_from_yaml():
     yml = plan.to_yaml()
     plan2 = Plan.from_yaml(yml, "pandas")
     assert plan == plan2
+
+
+def test_graph_plan_rule_without_named_output():
+    plan = Plan()
+    plan.add_rule(SortRule(['A'], named_input="input", named_output="sorted_data"))
+    plan.add_rule(ProjectRule(['A', 'B'], named_input="sorted_data", named_output="projected_data"))
+    with pytest.raises(InvalidPlanError) as exc:
+        plan.add_rule(RenameRule({'A': 'AA', 'B': 'BB'}, named_input="projected_data"))
+    assert "Mixing of rules taking named inputs and rules with no named inputs is not supported." in str(exc.value)